skip async intent resolutions with outstanding write
this makes handleSkippedIntents a little smarter: when
there is a write command gated for the intent's key,
the intent is ignored. This makes sense since there is
no "inconsistent write": a write will always have to
deal with the intent (either because it's an intent-
related command, or because it cares about the value).
In particular, repeated calls to handleSkippedIntents
will only ever have one active goroutine for an intent.