allow GC of undeleted inline keys
when scrubbing the Txn table, it would be awkward
to have to delete them individually, only to then
GC them after.

I suspect we want something similar for meta keys
when we implement proper TTLs (i.e. auto-clearing
keys written after X units of time), but we can
do that if we ever get there.