gc the transaction table
this is step two (and not the last) for safely garbage collecting
transaction entries (see #2062).

EndTransaction calls now store the non-local portion of the intents
in their transaction record (if there are any; otherwise the record
is instantly deleted; this is not new in this change).

on each run of the GC queue for a given range, its transaction table
is scanned and the following actions taken

* intents referenced from any "old" transaction records are resolved
  asynchronously
* old pending transactions are pushed (which will succeed), effectively
  aborting them
* old aborted transactions are added to the GC request.

TODO:
* necessary to send extra GC request?
* unregister intents from their Txn's entries when they've been resolved
  (currently they are never removed, so each GC run will waste work resolving
  a long gone intent every time and can never GC Txns with non-local intents)