batch expansion at (*Store.ExecuteCmd)
updated ExecuteCmd so that it properly unrolls a Batch.
Note that it has to do hacky things with `Key` and `EndKey`
to make sure they are truncated appropriately if the `BatchRequest`
is (after all, that's all `DistSender` adapts for cross-range ops).
Other than that, we mostly propagate the whole header from the Batch
and the Txn (if any) from each response to the next request.

`TxnCoordSender` still unrolls Batches and `(*DistSender).Send` still
wraps each individual request in a new Batch.

Some store tests use batches "properly", and so does `resolveWriteIntentError`,
so the new functionality is actually tested and it took a bit to get all
tests to pass again.

Note also that the first simplification was made: the test sender in ./storage
lost its fake batch handling.