Filtering in index-joins should only happen at the table.
When performing an index-join, we should only filter rows at the table
scan since the index-scan will not necessarily have all of the columns
needed by the filter.

Fixed a bug where index-join iteration was exiting early when the
index-scan returned a batch of rows which were then all filtered by the
table scan.