Handle an additional simplification case.
When simplifying the sub-expressions of an AND expression, the
simplification might create new AND expressions (or propagate the AND to
the top-level) allowing additional simplification opportunities.

Fixed a bug during AND/OR simplification where removal of an expression
from a list was being performed improperly causing the wrong element to
be removed an a nil expression to end up in the list.