sql: implement schema change lease
Once schema changes are scheduled to be executed through mutations,
a goroutine on the cluster acquires the schema change lease and executes
schema changes on the table. Any goroutine can acquire the lease.
In the future, the goroutine that executes the schema change command
(RENAME, ALTER, etc) will try to acquire the lease immediately
after scheduling a schema change  on the table descriptor.
Additionally, an assigned goroutine on every node will wake up on
a gossip notifications and attempt to acquire the lease if it
sees outstanding up_version bit or mutations on a table descriptor.