Detect on-disk state when creating replicas from raft messages.
This is a rare situation that occurs when a raft message arrives while a
replica is being garbage collected. The GC is aborted but the new
replica was being created without knowledge of the on-disk state.

This approach was abandoned in favor of #3709 in order to simplify
rather than add complexity to this rare case, but we may reconsider in
the future in order to move the expensive Destroy call out of Store.mu.