roachpb: Remove Method type
The roachpb.Method type is sometimes a convenient shorthand for
referring to different kinds of roachpb.Requests, but where it matters
we always use type switches. The Method of a request has occasionally
diverged from its type, leading to problems as in #11919.

Additionally, as a part of cleaning up #10084, I'm about to introduce a
new type storage.Command which makes roachpb.Method feel redundant.